+++ anybox/recipe/openerp/tests/test_server.py
@@ -266,7 +266,9 @@ class TestServer(RecipeTestCase):
         self.recipe.missing_deps_instructions[softreq] = (
             "This is an expected condition in this test.")
         self.recipe.soft_requirements = (softreq,)
-        self.recipe.b_options['offline'] = 'true'  # not enough, sadly
+
+        # offline won't be  enough, sadly for zc.buildout < 2.0
+        self.recipe.b_options['offline'] = 'true'
         self.unreachable_distributions.add(softreq)
 
         self.install_scripts(extra_requirements=(softreq,))